requirements.txt
behave
behavex
playwright
python-dotenv
requests
Faker
names